SearchEngine-Backend/searchrequirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
gunicorn==20.1.0
marshmallow==3.19.0
mysql-connector-python==8.0.31
pandas==1.3.5
yake==0.4.8
scikit-learn==1.0.2
nltk==3.8.1
txtai==5.2.0
python-decouple==3.8